changed type returned by Detector.

+++ TestSmellDetector/src/detector/Detector.java
@@ -8,6 +8,6 @@ package detector;
 public interface Detector extends Runnable{
 
 	public void run();
-	public int analyze();
+	public double analyze();
 	
 }
